requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.19.0
nibabel>=3.2.0
nilearn>=0.9.0
scipy>=1.5.0
scikit-learn>=0.23.0
tqdm>=4.50.0
pyyaml>=5.4
h5py>=3.0.0
matplotlib>=3.3.0  # Optional: for visualization
tensorboard>=2.4.0  # Optional: for logging
